fix crappy warnings
git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@20138 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ mount_slash/mount_slash_int.c
@@ -2484,7 +2484,7 @@ msl_io(struct pscfs_req *pfr, struct msl_fhent *mfh, char *buf,
 				rc = -EIO;
 				break;
 			}
-			PFL_GOTOERR(out, rc);
+			PFL_GOTOERR(out, (int)rc);
 		}
 
 		rc = msl_bmap_lease_tryext(b, NULL, 1);
@@ -2523,7 +2523,7 @@ msl_io(struct pscfs_req *pfr, struct msl_fhent *mfh, char *buf,
 				if (rc)
 					goto retry_bmap;
 
-				PFL_GOTOERR(out, rc = -EIO);
+				PFL_GOTOERR(out, (int)(rc = -EIO));
 			}
 		}
 		BMAP_CLI_BUMP_TIMEO(b);
@@ -2563,7 +2563,7 @@ msl_io(struct pscfs_req *pfr, struct msl_fhent *mfh, char *buf,
 					q->mfsrq_biorq[i] = NULL;
 					goto restart;
 				}
-				PFL_GOTOERR(out, rc = -EIO);
+				PFL_GOTOERR(out, (int)(rc = -EIO));
 			}
 		} else {
 			/*
@@ -2610,7 +2610,7 @@ msl_io(struct pscfs_req *pfr, struct msl_fhent *mfh, char *buf,
 				    (r->biorq_flags & BIORQ_READ) ?
 				    BIORQ_READFAIL : BIORQ_RBWFAIL;
 
-				PFL_GOTOERR(out, rc = -EIO);
+				PFL_GOTOERR(out, (int)(rc = -EIO));
 			}
 
 			if (rw == SL_READ)
